requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-secret-manager==2.16.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.22.3
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pytest==7.3.0
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
